drf_hls_video/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==4.4.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ffmpeg==1.4
flake8==6.0.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.1.0
Markdown==3.4.1
mccabe==0.7.0
moviepy==1.0.3
numpy==1.23.5
Pillow==9.3.0
proglog==0.1.10
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-ffmpeg-video-streaming==0.1.15
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.13
zipp==3.11.0